Did some refactoring and created a final branch to work on
