--- Main.py
@@ -1,9 +1,0 @@
-'''
-This is the main program of the system
-'''
-
-import datetime
-import configInterface
-
-
-
