+++ config.cfg
@@ -1,9 +1,9 @@
 [default]
-password = root
+password = tvzoiyy21611
 host = localhost
 user = root
-name = RemoteDB
-port = 8889
+name = hejhej
+port = 3306
 
 [createremote]
 name = guitest
@@ -30,14 +30,13 @@ port = 3306
 user = pimakers
 
 [channels]
-1 = ['101', 'Adams', 'C', '0.2']
-2 = ['102', 'Nurres', 'K', '0.1']
+1 = ['101', 'asd', 'asd', '234']
 
 [latestsession]
-start = 2017-05-18 13:34:29
-startfractions = 679925
-localdatabase = {'password': 'root', 'host': 'localhost', 'user': 'root', 'name': 'RemoteDB', 'port': '8889'}
-localsessionid = 1
-timeintervall = 2.0
-end = 2017-05-18 13:35:17
+timeintervall = 1.0
+startfractions = 838627
+localdatabase = {'password': 'tvzoiyy21611', 'user': 'root', 'port': '3306', 'host': 'localhost', 'name': 'hejhej'}
+end = 2017-05-22 22:04:57
+start = 2017-05-22 22:04:53
+localsessionid = 102
